Create broker.js
Simple broker script for the stockmarket in Bitburner. Scans the market, picks the #1 spot regarding forecast and buys in for the amount of 90% of player money. Then it waits and scans every 6 seconds (the time the stock market also refreshes) and if the forecast falls below the magic threshold of 0.5 (where the + becomes the - in the game, or 0.5 via the TIX API) it immediately sells all the bought stonks. After that it just starts again.

A few things I'm not too satisfied with: 
- If you kill the script you still have to sell your stock yourself.
- Sometimes the bought stock doesn't stay in the #1 spot, but the script doesn't sell, because it doesn't fall below the threshold. (on the other hand, that's how I played the stock market myself, so that's just a possible improvement.)
- If you sell by hand while the script runs, it doesn't realize it. I haven't found a way to make it realize that there are no longer stonks to sell. There probably is, but I'm still new to this.